fix bug turning supras back into homogeneous
Fix a bug whereby already heterogeneous supracontexts could become
homogeneous by having a deterministic subcontext added to them.
Add a regression test.

Also re-write `add(sub)` so that it only changes `outcome` when it has
to. This takes more logic, but it makes it clearer how heterogeneity
works.

+++ src/test/java/weka/classifiers/lazy/AM/TestUtils.java
@@ -231,7 +231,7 @@ public class TestUtils {
 			Subcontext sub = new Subcontext(label);
 
 			// parse outcome
-			int outcome;
+			double outcome;
 			if (subComponents[1].equals(AMUtils.NONDETERMINISTIC_STRING))
 				outcome = AMUtils.NONDETERMINISTIC;
 			else {
